snap-api: integrated general snap handling functions
