--- packages/galactica-dapp/src/types/snap.ts
@@ -1,8 +1,0 @@
-export type GetSnapsResponse = Record<string, Snap>;
-
-export type Snap = {
-  permissionName: string;
-  id: string;
-  version: string;
-  initialPermissions: Record<string, unknown>;
-};
