+++ packages/galactica-dapp/src/utils/button.ts
@@ -1,4 +1,5 @@
-import { isLocalSnap } from './snap';
+import { isLocalSnap } from '@galactica-net/snap-api';
+
 import { Snap } from '../types';
 
 export const shouldDisplayReconnectButton = (installedSnap?: Snap) =>
