+++ packages/galactica-dapp/src/utils/index.ts
@@ -1,5 +1,4 @@
 export * from './metamask';
-export * from './snap';
 export * from './theme';
 export * from './localStorage';
 export * from './button';
