+++ packages/galactica-passport-poc/src/pages/index.tsx
@@ -2,8 +2,6 @@ import { useContext } from 'react';
 import styled from 'styled-components';
 import { MetamaskActions, MetaMaskContext } from '../../../galactica-dapp/src/hooks';
 import {
-  connectSnap,
-  getSnap,
   shouldDisplayReconnectButton,
   queryVerificationSBTs,
   formatVerificationSBTs,
@@ -34,6 +32,8 @@ import {
   ZkCertStandard,
   ZkCertProof,
   HolderCommitmentData,
+  connectSnap,
+  getSnap,
 } from '@galactica-net/snap-api';
 import { zkKYCAgeProofPublicInputDescriptions, zkKYCPublicInputDescriptions } from '../../../galactica-dapp/src/config/snap';
