+++ packages/galactica-passport-poc/src/utils/index.ts
@@ -1,5 +1,4 @@
 export * from '../../../galactica-dapp/src/utils/metamask';
-export * from '../../../galactica-dapp/src/utils/snap';
 export * from '../../../galactica-dapp/src/utils/theme';
 export * from '../../../galactica-dapp/src/utils/localStorage';
 export * from '../../../galactica-dapp/src/utils/button';
