+++ packages/snap-api/.eslintrc.js
@@ -3,7 +3,11 @@ module.exports = {
 
   overrides: [
     {
-      files: ['src/utils/invoke-snap.ts'],
+      files: [
+        'src/utils/invoke-snap.ts',
+        'src/api/snap.ts',
+        'src/types/index.d.ts',
+      ],
       rules: {
         'no-restricted-globals': 0, // using window is ok in the browser, not sure how to disable the rule only for window.
       },
