--- packages/snap-api/src/global.d.ts
@@ -1,3 +1,0 @@
-type Window = {
-  ethereum: any;
-};
