+++ packages/snap-api/src/types/index.d.ts
@@ -1,0 +1,10 @@
+export {};
+
+/*
+ * Window type extension to support ethereum
+ */
+declare global {
+  type Window = typeof window & {
+    ethereum: any;
+  };
+}
