+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/Galactica-corp/galactica-snap.git"
   },
   "source": {
-    "shasum": "EW0wwo8+fyRwV0EN5u++jrkqkqT3Dj0g9PhNnMa7sYQ=",
+    "shasum": "IwXt1i3CR/iLfXX4xz/Mxa1XCf1KIJACoPlg1nyFqnk=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
